fix(agent-daemon): rename module path in files newly merged from main
Apply the same mechanical rename (agentactivity/daemon -> agent/daemon)
to runtime files added on main after the branch point
(claude_sdk_goal.go, claude_sdk_goal_lifecycle_test.go,
turn_lifecycle_stamp.go), which broke CI on the merge result.
