+++ packages/agent/daemon/runtime/turn_lifecycle_stamp.go
@@ -3,7 +3,7 @@ package agentruntime
 import (
 	"strings"
 
-	activityshared "github.com/tutti-os/tutti/packages/agentactivity/daemon/activity/events"
+	activityshared "github.com/tutti-os/tutti/packages/agent/daemon/activity/events"
 )
 
 // stampAdapterTurnLifecycleEvents stamps an adapter-origin TurnLifecycle
